[FEATURES] 📈 Added measurements

+++ app/src/main/java/com/rodrigo/deeplarva/domain/entity/Picture.kt
@@ -39,5 +39,5 @@ data class Picture(
     val count: Int,
 
     @ColumnInfo(name = "time")
-    val time: Int
+    val time: Long
 )
